converted to page and rewrote configuration panel to a page to avoid buddy behavior.  Added commenting.  Still come weird behavior with cw721 ids and names when manually adding cw721.  Removed all unnecessary logging and files.
