+++ src/lib/graphql/hooks/app/useGetApp.tsx
@@ -1,10 +1,8 @@
 import {
-    QueryADOAppAddress as QueryAppAddress,
     QueryADOAppAddressText as QueryAppAddressText,
     QueryADOAppAddressResponse as QueryAppAddressResponse,
-    QueryCW721AddressByName as QueryCW721Address,
-    QueryCW721AddressByNameText as QueryCW721AddressText,
-    QueryCW721AddressByNameResponse as QueryCW721AddressResponse,
+    QueryCW721AddressByNameText,
+    QueryCW721AddressByNameResponse,
     QueryCW721TokensResponse as QueryCW721TokensResponse,
     QueryCW721TokensText as QueryCW721TokensText,
     QueryCW721AuctionResponse,
@@ -14,10 +12,12 @@ import {
     QueryChainDataResponse,
     QueryChainDataText,
     QueryChainDenominationResponse,
-    QueryChainDenominationText
+    QueryChainDenominationText,
+    QueryCW721AddressResponse,
+    QueryCW721AddressText
   } from "./customQueryModule";
-  import { ApolloClient, InMemoryCache, QueryResult, useQuery } from "@apollo/client";
-  import { useEffect, useMemo, useState } from "react";
+  import { ApolloClient, InMemoryCache } from "@apollo/client";
+  import { useEffect, useState } from "react";
   
   const endpoint = process.env.NEXT_PUBLIC_GRAPHQL_ENDPOINT;
   const client = new ApolloClient({ uri: endpoint, cache: new InMemoryCache() });
@@ -29,6 +29,7 @@ import {
   }
   
   
+// custom hook to validate the ADO application address and build the ADO application object
 const useGetApp =  (appAddress: string) => {
     
     const [error, setError] = useState<string>();
@@ -70,129 +71,141 @@ const useGetApp =  (appAddress: string) => {
 
 const getConfig = async (appAddress: string)=>{
 
+   // First query to see if the application address is valid.
    const appData = await client.query<QueryAppAddressResponse>({
     query: QueryAppAddressText,
     variables:{appAddress}
    });
 
    
-   //let's get the chain data
+   //let's get the chain data 
    const allChainData = await client.query<QueryChainDataResponse>({
     query: QueryChainDataText
    });
  
+   // get the correct chain ID based on the address prefix
    const chainResponse = await allChainData.data.chainConfigs.allConfigs.find(chain => appAddress.startsWith(chain.addressPrefix));
-   
    const chain = chainResponse.chainId;
- 
    appData.chainId = chain;
 
-   //now let's get the chain demonination
+   //now let's get the chain token demonination
    const allDenomData = await client.query<QueryChainDenominationResponse>({
     query: QueryChainDenominationText
    });
-
    const denominationResponse = await allDenomData.data.keplrConfigs.allConfigs.find(denom => (denom.chainId === chain));
-
    const denomination = denominationResponse?.currencies[0].coinDenom;
-   
    appData.coinDenom = denomination;
    
- const addressPromises =   appData.data.ADO.app.components.filter((component) => component.ado_type === "cw721").map(async (component)=>{
-    const  componentAddress  = await client.query<QueryCW721AddressResponse>({
-        query: QueryCW721AddressText,
-        variables: { appAddress, collectionName: component.name },
-      });
-    return{
-        name: component.name,
-        address: componentAddress.data.ADO.app.getAddress
-    }
-   })
-
- // get all the CW721 Addresses from our cw721 filtered, mapped query
- const allCW721 = await Promise.all(addressPromises);
-
-
- 
-
- const tokenPromises = allCW721.map( async (cw721)=>{
-    //get CW721 Token IDs
-
-       const allTokens = await client.query<QueryCW721TokensResponse>({
-        query: QueryCW721TokensText,
-        variables: { cw721Address: cw721.address },
-
+   // filter the objects by only cw721 and go get their addresses
+   const addressPromises =   appData.data.ADO.app.components.filter((component) => component.ado_type === "cw721").map(async (component)=>{
+       const  componentAddress  = await client.query<QueryCW721AddressByNameResponse>({
+            query: QueryCW721AddressByNameText,
+            variables: { appAddress, collectionName: component.name },
         });
-      
-    return{
-        id:cw721.name,
-        contractAddress: cw721.address,
-        tokens: [...allTokens.data.ADO.cw721.allTokens]
-    }
-
- });
-
- let allTokens = await Promise.all(tokenPromises);
-
-
- 
-
-
- allTokens = await Promise.all(
-    allTokens.map(async cw721 => {
-      let auctionObj: AuctionObject | null = null;
-      let index = 0;
-  
-      do {
-        auctionObj = await checkAuctionMarketInfo(cw721.contractAddress, cw721.tokens[index]);
-        index++;
-      } while (!auctionObj && index < cw721.tokens.length);
-  
-      // Update the cw721 object with the auction object
-      if (auctionObj) {
-
-        if (auctionObj.adoType === 'auction') {
-            return {
-            ...cw721,
-            auctionAddress: auctionObj.auctionAddress,
-            };
-        } else if (auctionObj.adoType === 'market'){
-            return {
-            ...cw721,
-            marketAddress: auctionObj.auctionAddress,
-            };
+        return{
+            name: component.name,
+            address: componentAddress.data.ADO.app.getAddress
         }
-      } else {
-        return cw721;
-      }
     })
-  );
 
-  allTokens[0].featured = true;
+    // get all the CW721 Addresses from our cw721 filtered, mapped query
+    const allCW721 = await Promise.all(addressPromises);
+
+    // get all of the tokens for each of the cw721 objects using the cw721 address
+    const tokenPromises = allCW721.map( async (cw721)=>{
+        //get CW721 Token IDs
+        const allTokens = await client.query<QueryCW721TokensResponse>({
+            query: QueryCW721TokensText,
+            variables: { cw721Address: cw721.address },
+
+            });
+        
+        return{
+            id:cw721.name,
+            contractAddress: cw721.address,
+            tokens: [...allTokens.data.ADO.cw721.allTokens]
+        }
+    });
+
+ 
+    // get cw721 objects with new address info
+    let allTokens = await Promise.all(tokenPromises);
+
+    //get the names for the labels and stubs.
+    allTokens = await Promise.all(
+        allTokens.map(async cw721=>{
+        
+            const cw721Data = await client.query<QueryCW721AddressResponse>({
+                query: QueryCW721AddressText,
+                variables:{cw721Address: cw721.contractAddress}
+            });
+        
+
+            return{
+                ...cw721,
+                name: cw721Data?.data.cw721.contractInfo.name,
+                stubLink: cw721Data?.data.cw721.contractInfo.name.replace(/\s+/g, '-').toLowerCase(),
+            }
+        })
+    );
+
+ 
+    // get the auction or market objects.
+    allTokens = await Promise.all(
+        allTokens.map(async cw721 => {
+            let auctionObj: AuctionObject | null = null;
+            let index = 0;
+
+            do {
+            auctionObj = await checkAuctionMarketInfo(cw721.contractAddress, cw721.tokens[index]);
+            index++;
+            } while (!auctionObj && index < cw721.tokens.length);
+
+            // Update the cw721 object with the auction object
+            if (auctionObj) {
+
+                if (auctionObj.adoType === 'auction') {
+                    return {
+                    ...cw721,
+                    auctionAddress: auctionObj.auctionAddress,
+                    };
+                } else if (auctionObj.adoType === 'market'){
+                    return {
+                    ...cw721,
+                    marketAddress: auctionObj.auctionAddress,
+                    };
+                }
+            } else {
+                return cw721;
+            }
+        })
+    );
+    
+    // set the first cw721 collection to featured
+    allTokens[0].featured = true;
+ 
+   
+
  
 
- const { components, ...appConfig} = appData.data.ADO.app;
-
- 
-
- const config = {
-    name: appData.data.ADO.app.config.name,
-    chainId: chain,
-    coinDenom: denomination,
-    appAddress: appAddress,
-    collections:[ ...allTokens],
-    featured:{
-        collectionId: allTokens[0].id,
-        tokenId: allTokens[0].tokens[0]
-    }
+    const config = {
+        name: appData.data.ADO.app.config.name,
+        chainId: chain,
+        coinDenom: denomination,
+        appAddress: appAddress,
+        collections:[ ...allTokens],
+        featured:{
+            collectionId: allTokens[0].id,
+            tokenId: allTokens[0].tokens[0]
+        }
 
 
- };
+    };
 
 
 
-
- return config;
+    console.log(config);
+    return config;
 
 }
 
@@ -203,6 +216,7 @@ type AuctionObject = {
   };
 
 
+// function to get auction or market information for the auction/market address.
 async function checkAuctionMarketInfo (cw721Address: string, tokenId: string): Promise<AuctionObject> {
 
     //get Auction Address if it exists.
@@ -229,10 +243,7 @@ async function checkAuctionMarketInfo (cw721Address: string, tokenId: string): P
     else {
         throw new Error("auctionAddress not found");
       }
-    //const { address, type } = auctionObj.data.ADO.auctionMarketInfo;
-    
-
-
+  
 }
