+++ src/lib/graphql/hooks/collection/index.ts
@@ -1,4 +1,3 @@
 export { default as useGetCollection } from './useGetCollection'
-export { default as useGetCollectionByAddress} from './useGetCollectionByAddress'
 export { default as useGetTokens } from './useGetTokens'
 export * from './useGetToken'
