+++ src/modules/config/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Config } from './Config'
